moved context folder in folder structure / updated imports
