+++ src/pages/Login/index.tsx
@@ -5,7 +5,7 @@ import { login_InitVals } from "./utils/utils_loginForm";
 import { login_ValidationSchema } from "./utils/utils_loginForm";
 import { LOGIN_ENDPOINT } from "../../utils/constants/apiEndpoints";
 import { useLocation, useNavigate } from "react-router-dom";
-import { useAuth } from "../../utils/context/auth";
+import { useAuth } from "../../context/auth";
 
 
 const LoginPage = () => {
